Test the workspace app machinery under the license it ships behind
Workspace apps are a Pro feature, and the machinery tests ran on the free
version. They passed because restoring a saved tab and waking one are not
gated, which is a bug rather than a behaviour to lean on: every way of
acquiring a workspace app checks the license, and nothing checks it on the way
back in from config. A trial that ends is enough to reach it, with no config
edited by hand.

A test standing on that would have failed the day someone closed the hole, and
told them they had broken something when they had fixed it. So the six
machinery tests move to the Pro file, where two accounts also make session
partitioning assertable, and the count of restored views is now taken per URL
rather than as a total — which says "not materialized twice" directly instead
of by arithmetic over however many accounts there are.

What is left on the free version is the gate itself. Its account has a single
tab, because the saved tabs that would make a second are the Pro feature, and
`getVerticalTabsWidth` gives a one-tab account no strip at all — so the
launcher is placed in the sidebar, which keeps the strip for good. Seeding
saved tabs would have drawn the strip too, by standing on the same bug.

The hole is written up in the todo, with the trial path that reaches it and
the asymmetry that suggests it was never intended: a downgrade does slice the
account list back to one, so that gate survives it and this one does not.
